[#4] Repackage IUs to SVNKit 1.10 specific names
ru.arsysop.svn.connector -> ru.arsysop.svn.connector.svnkit1_10
and so on
