+++ bundles/ru.arsysop.svn.connector.svnkit1_10/META-INF/MANIFEST.MF
@@ -1,0 +1,16 @@
+Manifest-Version: 1.0
+Automatic-Module-Name: ru.arsysop.svn.connector.svnkit1_10
+Bundle-ManifestVersion: 2
+Bundle-SymbolicName: ru.arsysop.svn.connector.svnkit1_10;singleton:=true
+Bundle-Version: 0.1.0.qualifier
+Bundle-Name: %Bundle-Name
+Bundle-Vendor: %Bundle-Vendor
+Bundle-RequiredExecutionEnvironment: JavaSE-17
+Import-Package: org.apache.subversion.javahl.types;version="1.10.11",
+ org.tmatesoft.sqljet.core;version="1.1.15",
+ org.tmatesoft.svn.core.javahl17;version="1.10.11",
+ org.tmatesoft.svn.util;version="1.10.11"
+Require-Bundle: org.eclipse.core.runtime;bundle-version="3.18.0",
+ org.eclipse.team.svn.core;bundle-version="5.0.0"
+Bundle-ActivationPolicy: lazy
+Export-Package: sun.security.x509
