+++ bundles/ru.arsysop.svn.connector.svnkit1_10/build.properties
