+++ bundles/ru.arsysop.svn.connector.svnkit1_10/plugin.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?eclipse version="3.0"?>
+<!--
+# Copyright (c) 2023, 2024 ArSysOp
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# SPDX-License-Identifier: Apache-2.0
+#
+# Contributors:
+#     ArSysOp - initial API and implementation 
+-->
+
+<plugin>
+
+	<extension point="org.eclipse.team.svn.core.svnconnector">
+		<svnconnector class="ru.arsysop.svn.connector.internal.svnkit1_10.SvnKit1_10ConnectorFactory"/>
+	</extension>
+	
+</plugin>
