+++ features/ru.arsysop.svn.connectors.svnkit1_10.feature/build.properties
