+++ features/ru.arsysop.svn.connectors.svnkit1_10.feature/feature.xml
@@ -1,0 +1,46 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+# Copyright (c) 2023, 2024 ArSysOp
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# SPDX-License-Identifier: Apache-2.0
+#
+# Contributors:
+#     ArSysOp - initial API and implementation 
+-->
+<feature
+      id="ru.arsysop.svn.connectors.svnkit1_10.feature"
+      label="%feature.name"
+      version="0.1.0.qualifier"
+      provider-name="%feature.vendor"
+      license-feature="ru.arsysop.svn.license.feature"
+      license-feature-version="0.1.0.qualifier">
+
+   <description>
+      %feature.description
+   </description>
+
+   <copyright>
+      %copyright
+   </copyright>
+
+   <license url="%licenseURL">
+      %license
+   </license>
+
+   <plugin
+         id="ru.arsysop.svn.connector.svnkit1_10"
+         version="0.0.0"/>
+
+</feature>
