+++ tests/ru.arsysop.svn.connector.svnkit1_10.tests/.classpath
