+++ tests/ru.arsysop.svn.connector.svnkit1_10.tests/OSGI-INF/l10n/bundle.properties
