--- tests/ru.arsysop.svn.connector.tests/META-INF/MANIFEST.MF
@@ -1,15 +1,0 @@
-Manifest-Version: 1.0
-Automatic-Module-Name: ru.arsysop.svn.connector.tests
-Bundle-ManifestVersion: 2
-Bundle-SymbolicName: ru.arsysop.svn.connector.tests;singleton:=true
-Bundle-Version: 0.1.0.qualifier
-Bundle-ClassPath: .
-Bundle-Name: %Bundle-Name
-Bundle-Vendor: %Bundle-Vendor
-Bundle-RequiredExecutionEnvironment: JavaSE-17
-Bundle-ActivationPolicy: lazy
-Import-Package: org.junit;version="4.12.0",
- org.junit.jupiter.api;version="5.10.0",
- org.junit.jupiter.api.function;version="5.10.0",
- org.junit.jupiter.engine;version="5.10.0"
-Fragment-Host: ru.arsysop.svn.connector
